RUB-288 rust: split txcontext tests
