+++ clients/rust/crates/rubin-consensus/src/tests/mod.rs
@@ -625,4 +625,5 @@ mod parse_dedup;
 mod precompute;
 mod tx_parse;
 mod tx_validate_worker;
+mod txcontext;
 mod utxo_apply;
