+++ clients/rust/crates/rubin-consensus/src/txcontext.rs
@@ -266,569 +266,3 @@ pub fn build_tx_context(
 
     Ok(Some(TxContextBundle { base, continuing }))
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::core_ext::{CoreExtActiveProfile, CoreExtProfiles, CoreExtVerificationBinding};
-    use crate::tx::{Tx, TxInput, TxOutput};
-
-    fn core_ext_covdata(ext_id: u16, payload: &[u8]) -> Vec<u8> {
-        crate::core_ext::encode_core_ext_covenant_data(ext_id, payload)
-            .expect("CORE_EXT covenant_data encode")
-    }
-
-    fn static_profiles(entries: &[(u16, bool)]) -> CoreExtProfiles {
-        CoreExtProfiles {
-            active: entries
-                .iter()
-                .map(|(ext_id, tx_context_enabled)| CoreExtActiveProfile {
-                    ext_id: *ext_id,
-                    tx_context_enabled: *tx_context_enabled,
-                    allowed_suite_ids: vec![0x42],
-                    verification_binding: CoreExtVerificationBinding::VerifySigExtAccept,
-                    verify_sig_ext_tx_context_fn: None,
-                    binding_descriptor: b"accept".to_vec(),
-                    ext_payload_schema: b"schema".to_vec(),
-                })
-                .collect(),
-        }
-    }
-
-    #[test]
-    fn build_tx_context_output_ext_id_cache_rejects_malformed_output() {
-        let tx = Tx {
-            version: 1,
-            tx_kind: 0,
-            tx_nonce: 1,
-            inputs: vec![TxInput {
-                prev_txid: [0u8; 32],
-                prev_vout: 0,
-                script_sig: vec![],
-                sequence: 0,
-            }],
-            outputs: vec![TxOutput {
-                value: 10,
-                covenant_type: COV_TYPE_CORE_EXT,
-                covenant_data: vec![0x01],
-            }],
-            locktime: 0,
-            da_commit_core: None,
-            da_chunk_core: None,
-            witness: vec![],
-            da_payload: vec![],
-        };
-
-        let err = build_tx_context_output_ext_id_cache(&tx).unwrap_err();
-        assert_eq!(err.code, ErrorCode::TxErrCovenantTypeInvalid);
-    }
-
-    #[test]
-    fn txcontext_get_output_checked_rejects_missing_slot() {
-        let continuing = TxContextContinuing {
-            continuing_output_count: 1,
-            continuing_outputs: [None, None],
-        };
-
-        let err = continuing.get_output_checked(0).unwrap_err();
-        assert_eq!(err.code, ErrorCode::TxErrSigInvalid);
-        assert_eq!(err.msg, "txcontext continuing output missing");
-    }
-
-    #[test]
-    fn build_tx_context_returns_none_without_txcontext_enabled_inputs() {
-        let tx = Tx {
-            version: 1,
-            tx_kind: 0,
-            tx_nonce: 1,
-            inputs: vec![TxInput {
-                prev_txid: [0u8; 32],
-                prev_vout: 0,
-                script_sig: vec![],
-                sequence: 0,
-            }],
-            outputs: vec![TxOutput {
-                value: 33,
-                covenant_type: COV_TYPE_CORE_EXT,
-                covenant_data: core_ext_covdata(7, &[0xaa]),
-            }],
-            locktime: 0,
-            da_commit_core: None,
-            da_chunk_core: None,
-            witness: vec![],
-            da_payload: vec![],
-        };
-        let resolved_inputs = vec![UtxoEntry {
-            value: 50,
-            covenant_type: COV_TYPE_CORE_EXT,
-            covenant_data: core_ext_covdata(7, &[0xbb]),
-            creation_height: 0,
-            created_by_coinbase: false,
-        }];
-
-        let cache = build_tx_context_output_ext_id_cache(&tx).unwrap();
-        let bundle = build_tx_context(
-            &tx,
-            &resolved_inputs,
-            Some(&cache),
-            12,
-            &static_profiles(&[(7, false)]),
-        )
-        .unwrap();
-        assert!(bundle.is_none());
-    }
-
-    #[test]
-    fn collect_txcontext_ext_ids_skips_disabled_profiles_and_deduplicates() {
-        let resolved_inputs = vec![
-            UtxoEntry {
-                value: 11,
-                covenant_type: COV_TYPE_CORE_EXT,
-                covenant_data: core_ext_covdata(9, &[0x90]),
-                creation_height: 0,
-                created_by_coinbase: false,
-            },
-            UtxoEntry {
-                value: 12,
-                covenant_type: COV_TYPE_CORE_EXT,
-                covenant_data: core_ext_covdata(7, &[0x71]),
-                creation_height: 0,
-                created_by_coinbase: false,
-            },
-            UtxoEntry {
-                value: 13,
-                covenant_type: COV_TYPE_CORE_EXT,
-                covenant_data: core_ext_covdata(7, &[0x72]),
-                creation_height: 0,
-                created_by_coinbase: false,
-            },
-            UtxoEntry {
-                value: 14,
-                covenant_type: crate::constants::COV_TYPE_P2PK,
-                covenant_data: vec![],
-                creation_height: 0,
-                created_by_coinbase: false,
-            },
-        ];
-
-        let ext_ids =
-            collect_txcontext_ext_ids(&resolved_inputs, &static_profiles(&[(7, true), (9, false)]))
-                .expect("collect ext_ids");
-        assert_eq!(ext_ids, vec![7]);
-    }
-
-    #[test]
-    fn collect_txcontext_ext_ids_rejects_duplicate_active_profiles() {
-        let resolved_inputs = vec![UtxoEntry {
-            value: 11,
-            covenant_type: COV_TYPE_CORE_EXT,
-            covenant_data: core_ext_covdata(7, &[0x71]),
-            creation_height: 0,
-            created_by_coinbase: false,
-        }];
-        let profiles = CoreExtProfiles {
-            active: vec![
-                CoreExtActiveProfile {
-                    ext_id: 7,
-                    tx_context_enabled: true,
-                    allowed_suite_ids: vec![0x42],
-                    verification_binding: CoreExtVerificationBinding::VerifySigExtAccept,
-                    verify_sig_ext_tx_context_fn: None,
-                    binding_descriptor: b"accept-a".to_vec(),
-                    ext_payload_schema: b"schema-a".to_vec(),
-                },
-                CoreExtActiveProfile {
-                    ext_id: 7,
-                    tx_context_enabled: true,
-                    allowed_suite_ids: vec![0x43],
-                    verification_binding: CoreExtVerificationBinding::VerifySigExtAccept,
-                    verify_sig_ext_tx_context_fn: None,
-                    binding_descriptor: b"accept-b".to_vec(),
-                    ext_payload_schema: b"schema-b".to_vec(),
-                },
-            ],
-        };
-
-        let err = collect_txcontext_ext_ids(&resolved_inputs, &profiles).unwrap_err();
-        assert_eq!(err.code, ErrorCode::TxErrCovenantTypeInvalid);
-        assert_eq!(err.msg, "CORE_EXT multiple ACTIVE profiles for ext_id");
-    }
-
-    #[test]
-    fn build_tx_context_requires_output_cache_when_enabled() {
-        let tx = Tx {
-            version: 1,
-            tx_kind: 0,
-            tx_nonce: 1,
-            inputs: vec![TxInput {
-                prev_txid: [0u8; 32],
-                prev_vout: 0,
-                script_sig: vec![],
-                sequence: 0,
-            }],
-            outputs: vec![TxOutput {
-                value: 33,
-                covenant_type: COV_TYPE_CORE_EXT,
-                covenant_data: core_ext_covdata(7, &[]),
-            }],
-            locktime: 0,
-            da_commit_core: None,
-            da_chunk_core: None,
-            witness: vec![],
-            da_payload: vec![],
-        };
-        let resolved_inputs = vec![UtxoEntry {
-            value: 50,
-            covenant_type: COV_TYPE_CORE_EXT,
-            covenant_data: core_ext_covdata(7, &[0xbb]),
-            creation_height: 0,
-            created_by_coinbase: false,
-        }];
-
-        let err = build_tx_context(
-            &tx,
-            &resolved_inputs,
-            None,
-            12,
-            &static_profiles(&[(7, true)]),
-        )
-        .unwrap_err();
-        assert_eq!(err.code, ErrorCode::TxErrCovenantTypeInvalid);
-    }
-
-    #[test]
-    fn build_tx_context_rejects_resolved_input_count_mismatch() {
-        let tx = Tx {
-            version: 1,
-            tx_kind: 0,
-            tx_nonce: 1,
-            inputs: vec![TxInput {
-                prev_txid: [0u8; 32],
-                prev_vout: 0,
-                script_sig: vec![],
-                sequence: 0,
-            }],
-            outputs: vec![TxOutput {
-                value: 33,
-                covenant_type: COV_TYPE_CORE_EXT,
-                covenant_data: core_ext_covdata(7, &[]),
-            }],
-            locktime: 0,
-            da_commit_core: None,
-            da_chunk_core: None,
-            witness: vec![],
-            da_payload: vec![],
-        };
-
-        let err = build_tx_context(
-            &tx,
-            &[],
-            Some(&BTreeMap::new()),
-            12,
-            &static_profiles(&[(7, true)]),
-        )
-        .unwrap_err();
-        assert_eq!(err.code, ErrorCode::TxErrParse);
-        assert_eq!(err.msg, "txcontext resolved input count mismatch");
-    }
-
-    #[test]
-    fn build_tx_context_keeps_enabled_ext_id_with_empty_continuations() {
-        let tx = Tx {
-            version: 1,
-            tx_kind: 0,
-            tx_nonce: 1,
-            inputs: vec![TxInput {
-                prev_txid: [0u8; 32],
-                prev_vout: 0,
-                script_sig: vec![],
-                sequence: 0,
-            }],
-            outputs: vec![TxOutput {
-                value: 33,
-                covenant_type: crate::constants::COV_TYPE_P2PK,
-                covenant_data: vec![0x01],
-            }],
-            locktime: 0,
-            da_commit_core: None,
-            da_chunk_core: None,
-            witness: vec![],
-            da_payload: vec![],
-        };
-        let resolved_inputs = vec![UtxoEntry {
-            value: 50,
-            covenant_type: COV_TYPE_CORE_EXT,
-            covenant_data: core_ext_covdata(7, &[0xbb]),
-            creation_height: 0,
-            created_by_coinbase: false,
-        }];
-
-        let cache = build_tx_context_output_ext_id_cache(&tx).expect("cache");
-        let bundle = build_tx_context(
-            &tx,
-            &resolved_inputs,
-            Some(&cache),
-            12,
-            &static_profiles(&[(7, true)]),
-        )
-        .unwrap()
-        .expect("bundle");
-
-        assert_eq!(bundle.sorted_ext_ids(), vec![7]);
-        let continuing = bundle.get_continuing(7).expect("ext 7");
-        assert_eq!(continuing.continuing_output_count, 0);
-        assert!(continuing.valid_outputs().is_empty());
-    }
-
-    #[test]
-    fn build_tx_context_builds_base_and_deterministic_continuations() {
-        let tx = Tx {
-            version: 1,
-            tx_kind: 0,
-            tx_nonce: 1,
-            inputs: vec![
-                TxInput {
-                    prev_txid: [1u8; 32],
-                    prev_vout: 0,
-                    script_sig: vec![],
-                    sequence: 0,
-                },
-                TxInput {
-                    prev_txid: [2u8; 32],
-                    prev_vout: 1,
-                    script_sig: vec![],
-                    sequence: 0,
-                },
-                TxInput {
-                    prev_txid: [3u8; 32],
-                    prev_vout: 2,
-                    script_sig: vec![],
-                    sequence: 0,
-                },
-            ],
-            outputs: vec![
-                TxOutput {
-                    value: 11,
-                    covenant_type: COV_TYPE_CORE_EXT,
-                    covenant_data: core_ext_covdata(7, &[0x07, 0x01]),
-                },
-                TxOutput {
-                    value: 12,
-                    covenant_type: crate::constants::COV_TYPE_P2PK,
-                    covenant_data: vec![0x01],
-                },
-                TxOutput {
-                    value: 13,
-                    covenant_type: COV_TYPE_CORE_EXT,
-                    covenant_data: core_ext_covdata(5, &[0x05, 0x01]),
-                },
-                TxOutput {
-                    value: 14,
-                    covenant_type: COV_TYPE_CORE_EXT,
-                    covenant_data: core_ext_covdata(7, &[0x07, 0x02]),
-                },
-                TxOutput {
-                    value: 15,
-                    covenant_type: COV_TYPE_CORE_EXT,
-                    covenant_data: core_ext_covdata(5, &[0x05, 0x02]),
-                },
-            ],
-            locktime: 0,
-            da_commit_core: None,
-            da_chunk_core: None,
-            witness: vec![],
-            da_payload: vec![],
-        };
-        let resolved_inputs = vec![
-            UtxoEntry {
-                value: 100,
-                covenant_type: COV_TYPE_CORE_EXT,
-                covenant_data: core_ext_covdata(7, &[0xaa]),
-                creation_height: 0,
-                created_by_coinbase: false,
-            },
-            UtxoEntry {
-                value: 200,
-                covenant_type: COV_TYPE_CORE_EXT,
-                covenant_data: core_ext_covdata(5, &[0xbb]),
-                creation_height: 0,
-                created_by_coinbase: false,
-            },
-            UtxoEntry {
-                value: 300,
-                covenant_type: crate::constants::COV_TYPE_P2PK,
-                covenant_data: vec![0x01],
-                creation_height: 0,
-                created_by_coinbase: false,
-            },
-        ];
-
-        let cache = build_tx_context_output_ext_id_cache(&tx).unwrap();
-        let bundle = build_tx_context(
-            &tx,
-            &resolved_inputs,
-            Some(&cache),
-            222,
-            &static_profiles(&[(5, true), (7, true)]),
-        )
-        .unwrap()
-        .expect("bundle");
-        assert_eq!(bundle.base.total_in, Uint128 { lo: 600, hi: 0 });
-        assert_eq!(bundle.base.total_out, Uint128 { lo: 65, hi: 0 });
-        assert_eq!(bundle.base.height, 222);
-        assert_eq!(bundle.sorted_ext_ids(), vec![5, 7]);
-
-        let ext5 = bundle.get_continuing(5).expect("ext 5");
-        assert_eq!(ext5.continuing_output_count, 2);
-        assert_eq!(ext5.get_output_checked(0).unwrap().value, 13);
-        assert_eq!(
-            ext5.get_output_checked(0).unwrap().ext_payload.as_ref(),
-            &[0x05, 0x01]
-        );
-        assert_eq!(ext5.get_output_checked(1).unwrap().value, 15);
-
-        let ext7 = bundle.get_continuing(7).expect("ext 7");
-        assert_eq!(ext7.continuing_output_count, 2);
-        assert_eq!(ext7.get_output_checked(0).unwrap().value, 11);
-        assert_eq!(ext7.get_output_checked(1).unwrap().value, 14);
-    }
-
-    #[test]
-    fn build_tx_context_preserves_empty_payload_as_empty_vec() {
-        let tx = Tx {
-            version: 1,
-            tx_kind: 0,
-            tx_nonce: 1,
-            inputs: vec![TxInput {
-                prev_txid: [0u8; 32],
-                prev_vout: 0,
-                script_sig: vec![],
-                sequence: 0,
-            }],
-            outputs: vec![TxOutput {
-                value: 33,
-                covenant_type: COV_TYPE_CORE_EXT,
-                covenant_data: core_ext_covdata(7, &[]),
-            }],
-            locktime: 0,
-            da_commit_core: None,
-            da_chunk_core: None,
-            witness: vec![],
-            da_payload: vec![],
-        };
-        let resolved_inputs = vec![UtxoEntry {
-            value: 50,
-            covenant_type: COV_TYPE_CORE_EXT,
-            covenant_data: core_ext_covdata(7, &[0xaa]),
-            creation_height: 0,
-            created_by_coinbase: false,
-        }];
-        let cache = build_tx_context_output_ext_id_cache(&tx).unwrap();
-        let bundle = build_tx_context(
-            &tx,
-            &resolved_inputs,
-            Some(&cache),
-            12,
-            &static_profiles(&[(7, true)]),
-        )
-        .unwrap()
-        .expect("bundle");
-        let output = bundle
-            .get_continuing(7)
-            .unwrap()
-            .get_output_checked(0)
-            .unwrap()
-            .ext_payload
-            .clone();
-        assert!(output.is_empty());
-    }
-
-    #[test]
-    fn build_tx_context_rejects_third_output_for_lowest_ext_id() {
-        let tx = Tx {
-            version: 1,
-            tx_kind: 0,
-            tx_nonce: 1,
-            inputs: vec![
-                TxInput {
-                    prev_txid: [1u8; 32],
-                    prev_vout: 0,
-                    script_sig: vec![],
-                    sequence: 0,
-                },
-                TxInput {
-                    prev_txid: [2u8; 32],
-                    prev_vout: 1,
-                    script_sig: vec![],
-                    sequence: 0,
-                },
-            ],
-            outputs: vec![
-                TxOutput {
-                    value: 1,
-                    covenant_type: COV_TYPE_CORE_EXT,
-                    covenant_data: core_ext_covdata(9, &[0x91]),
-                },
-                TxOutput {
-                    value: 2,
-                    covenant_type: COV_TYPE_CORE_EXT,
-                    covenant_data: core_ext_covdata(7, &[0x71]),
-                },
-                TxOutput {
-                    value: 3,
-                    covenant_type: COV_TYPE_CORE_EXT,
-                    covenant_data: core_ext_covdata(9, &[0x92]),
-                },
-                TxOutput {
-                    value: 4,
-                    covenant_type: COV_TYPE_CORE_EXT,
-                    covenant_data: core_ext_covdata(7, &[0x72]),
-                },
-                TxOutput {
-                    value: 5,
-                    covenant_type: COV_TYPE_CORE_EXT,
-                    covenant_data: core_ext_covdata(7, &[0x73]),
-                },
-                TxOutput {
-                    value: 6,
-                    covenant_type: COV_TYPE_CORE_EXT,
-                    covenant_data: core_ext_covdata(9, &[0x93]),
-                },
-            ],
-            locktime: 0,
-            da_commit_core: None,
-            da_chunk_core: None,
-            witness: vec![],
-            da_payload: vec![],
-        };
-        let resolved_inputs = vec![
-            UtxoEntry {
-                value: 100,
-                covenant_type: COV_TYPE_CORE_EXT,
-                covenant_data: core_ext_covdata(9, &[0xaa]),
-                creation_height: 0,
-                created_by_coinbase: false,
-            },
-            UtxoEntry {
-                value: 200,
-                covenant_type: COV_TYPE_CORE_EXT,
-                covenant_data: core_ext_covdata(7, &[0xbb]),
-                creation_height: 0,
-                created_by_coinbase: false,
-            },
-        ];
-        let cache = build_tx_context_output_ext_id_cache(&tx).unwrap();
-        let err = build_tx_context(
-            &tx,
-            &resolved_inputs,
-            Some(&cache),
-            12,
-            &static_profiles(&[(7, true), (9, true)]),
-        )
-        .unwrap_err();
-        assert_eq!(err.code, ErrorCode::TxErrCovenantTypeInvalid);
-        assert!(err
-            .to_string()
-            .contains(TXCONTEXT_TOO_MANY_CONTINUING_OUTPUTS));
-    }
-}
